propter-optimis-app/backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# HTTP and API
requests==2.31.0
urllib3==2.0.7

# Environment
python-decouple==3.8

# Validation
marshmallow==3.20.1

# Date/Time
python-dateutil==2.8.2

# Utilities
PyJWT==2.8.0
cryptography==41.0.7

# Development
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# OpenStar Lab Intelligence Packages
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# OpenStar Lab (these would be actual PyPI packages in production)
# openstarlab-core==1.0.0
# openstarlab-event-modeling==1.0.0
# openstarlab-preprocessing==1.0.0
# openstarlab-rlearn==1.0.0

# Additional ML/AI packages for intelligence processing
torch==2.1.1
transformers==4.35.2
opencv-python-headless==4.8.1.78
